# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 电商平台 API
  description: |
    3 人小组项目共用接口契约。修改接口请先改本文档，再实现前后端。
    业务术语见 docs/BUSINESS_GLOSSARY.md，状态机见 docs/domain/DOMAIN_MODEL.md
    管理后台见 docs/domain/ADMIN.md
  version: 0.2.0
servers:
  - url: http://localhost:8080/api
    description: 本地开发（端口按实际后端调整）
...
paths:
...
  /auth/admin/login:
    post:
      tags: [auth]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'

  # ========== Product (C端) ==========
...
  /users/addresses:
    get:
      tags: [user]
      summary: 当前用户收货地址列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags: [user]
      summary: 新增收货地址
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAddressRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /users/addresses/{addressId}:
    put:
      tags: [user]
      summary: 修改收货地址
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      tags: [user]
      summary: 删除收货地址
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
  /users/addresses/{addressId}/default:
    patch:
      tags: [user]
      summary: 设为默认地址
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'

  # ========== Order ==========
...
  /admin/dashboard/summary:
    get:
      tags: [admin]
      summary: 工作台待办统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardSummary'
  /admin/products/audits:
    get:
      tags: [admin]
      summary: 商品审核历史记录
      security:
        - bearerAuth: []
      parameters:
        - name: approved
          in: query
          schema:
            type: string
            enum: ['true', 'false']
          description: 筛选通过/驳回
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductAuditListResponse'
  /admin/products/pending:
    get:
      tags: [admin]
      summary: 待审核商品列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingProductListResponse'
  /admin/products/{spuId}/audit:
    post:
      tags: [admin]
      summary: 审核商品
      security:
        - bearerAuth: []
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRequest'
      responses:
        '200':
          description: OK
  /admin/orders:
    get:
      tags: [admin]
      summary: 全平台订单列表
      security:
        - bearerAuth: []
      parameters:
        - name: orderNo
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: merchantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
  /admin/orders/{orderId}:
    get:
      tags: [admin]
      summary: 订单详情
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
  /admin/after-sales:
    get:
      tags: [admin]
      summary: 待仲裁售后列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AfterSaleListResponse'
...
  /admin/merchants/applications:
    get:
      tags: [admin]
      summary: 商家入驻申请列表（可按状态筛选）
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            description: PENDING / APPROVED / REJECTED / ALL，多个用逗号分隔
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantApplicationListResponse'
  /admin/merchants/pending:
    get:
      tags: [admin]
      summary: 待审核商家列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingMerchant'
  /admin/merchants/{merchantId}/audit:
    post:
      tags: [admin]
      summary: 审核商家入驻
      security:
        - bearerAuth: []
      parameters:
        - name: merchantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRequest'
      responses:
        '200':
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AdminAuthResponse:
      type: object
      properties:
        token:
          type: string
        adminId:
          type: integer
        role:
          $ref: '#/components/schemas/AdminRole'
    AdminRole:
      type: string
      enum:
        - OPERATOR
        - CS_AGENT
    AuditRequest:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
        reason:
          type: string
...
    AdminDashboardSummary:
      type: object
      properties:
        pendingProductCount:
          type: integer
          description: 待审核商品数（PENDING_AUDIT）
        escalatedAfterSaleCount:
          type: integer
          description: 待仲裁售后数（ESCALATED）
        pendingMerchantCount:
          type: integer
          description: 待审核商家入驻数
        auditedProductCount:
          type: integer
          description: 累计已审核商品次数
        onShelfProductCount:
          type: integer
          description: 已上架商品数
        rejectedProductCount:
          type: integer
          description: 已驳回商品数
        totalOrderCount:
          type: integer
          description: 全平台订单总数
        pendingPaymentOrderCount:
          type: integer
          description: 待支付订单数
        recentPendingProducts:
          type: array
          items:
            $ref: '#/components/schemas/PendingProductSummary'
        recentPendingMerchants:
          type: array
          items:
            $ref: '#/components/schemas/PendingMerchant'
        recentEscalatedAfterSales:
          type: array
          items:
            $ref: '#/components/schemas/AfterSale'
    ProductAuditRecord:
      type: object
      properties:
        auditId:
          type: integer
        spuId:
          type: integer
        approved:
          type: boolean
        reason:
          type: string
        auditedAt:
          type: string
          format: date-time
        adminId:
          type: integer
        title:
          type: string
        shopName:
          type: string
        merchantId:
          type: integer
        mainImage:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductAuditListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/ProductAuditRecord'
    PendingProductSummary:
      type: object
      properties:
        spuId:
          type: integer
        title:
          type: string
        shopName:
          type: string
        submittedAt:
          type: string
          format: date-time
    AfterSaleType:
      type: string
      enum:
        - REFUND_ONLY
        - RETURN_REFUND
    AfterSaleStatus:
      type: string
      enum:
        - APPLIED
        - APPROVED
        - REJECTED
        - ESCALATED
        - RETURNING
        - REFUNDED
...
    AfterSale:
      type: object
      properties:
        afterSaleId:
          type: integer
        orderId:
          type: integer
        orderNo:
          type: string
        subOrderId:
          type: integer
        merchantId:
          type: integer
        userId:
          type: integer
        shopName:
          type: string
        type:
          $ref: '#/components/schemas/AfterSaleType'
        reason:
          type: string
        status:
          $ref: '#/components/schemas/AfterSaleStatus'
          description: ESCALATED 表示待平台仲裁
        appliedAt:
          type: string
          format: date-time
        merchantDeadline:
          type: string
          format: date-time
        auditReason:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              skuId:
                type: integer
              title:
                type: string
              price:
                type: number
              quantity:
                type: integer
    AfterSaleListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/AfterSale'
    PendingProductListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            type: object
            properties:
              spuId:
                type: integer
              title:
                type: string
              shopName:
                type: string
              merchantId:
                type: integer
              mainImage:
                type: string
              submittedAt:
                type: string
                format: date-time
    PendingMerchant:
      type: object
      properties:
        merchantId:
          type: integer
          description: 申请 ID
        shopName:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        appliedAt:
          type: string
          format: date-time
...
    MerchantApplication:
      allOf:
        - $ref: '#/components/schemas/PendingMerchant'
        - type: object
          properties:
            status:
              type: string
              enum: [PENDING, APPROVED, REJECTED]
            auditedAt:
              type: string
              format: date-time
            rejectReason:
              type: string
            approvedMerchantId:
              type: integer
            merchantUsername:
              type: string
              description: 审核通过后生成的商家登录账号
    MerchantApplicationListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/MerchantApplication'
    OrderListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            subOrders:
              type: array
              items:
                $ref: '#/components/schemas/SubOrder'
            addressSnapshot:
              type: object
              properties:
                receiverName:
                  type: string
                phone:
                  type: string
                fullAddress:
                  type: string
    SubOrder:
      type: object
      properties:
        subOrderId:
          type: integer
        merchantId:
          type: integer
        shopName:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        shipment:
          type: object
          properties:
            logisticsCompany:
              type: string
            trackingNo:
              type: string
            shippedAt:
              type: string
              format: date-time
    ProductStatus:
      type: string
      enum:
        - DRAFT
        - PENDING_AUDIT
        - ON_SHELF
        - REJECTED
        - OFF_SHELF
...
    Address:
      type: object
      properties:
        id:
          type: integer
        receiverName:
          type: string
        phone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        isDefault:
          type: boolean
    CreateAddressRequest:
      type: object
      required: [receiverName, phone, province, city, district, detail]
      properties:
        receiverName:
          type: string
        phone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        isDefault:
          type: boolean
    UpdateAddressRequest:
      type: object
      properties:
        receiverName:
          type: string
        phone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        isDefault:
          type: boolean
...
    OrderStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - PENDING_SHIPMENT
        - SHIPPED
        - COMPLETED
        - CANCELLED
        - REFUNDING
        - REFUNDED
    Order:
      type: object
      properties:
        orderId:
          type: integer
        orderNo:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: number
        items:
          type: array
          items:
            type: object
            properties:
              skuId:
                type: integer
              title:
                type: string
              price:
                type: number
              quantity:
                type: integer
        createdAt:
          type: string
          format: date-time