# packages/web-admin/src/api/admin.js
import client from './client';

export function login(username, password) {
  return client.post('/auth/admin/login', { username, password }).then((r) => r.data);
}

export function fetchPendingProducts(params) {
  return client.get('/admin/products/pending', { params }).then((r) => r.data);
}

export function fetchProductAudits(params) {
  return client.get('/admin/products/audits', { params }).then((r) => r.data);
}

export function fetchProductDetail(spuId) {
  return client.get(`/admin/products/${spuId}`).then((r) => r.data);
}

export function auditProduct(spuId, approved, reason) {
  return client.post(`/admin/products/${spuId}/audit`, { approved, reason }).then((r) => r.data);
}

export function fetchOrders(params) {
  return client.get('/admin/orders', { params }).then((r) => r.data);
}

export function fetchOrderDetail(orderId) {
  return client.get(`/admin/orders/${orderId}`).then((r) => r.data);
}

export function fetchDashboardSummary() {
  return client.get('/admin/dashboard/summary').then((r) => r.data);
}

export function fetchEscalatedAfterSales(params) {
  return client.get('/admin/after-sales', { params }).then((r) => r.data);
}

export function fetchPendingMerchants() {
  return client.get('/admin/merchants/pending').then((r) => r.data);
}

export function fetchMerchantApplications(params) {
  return client.get('/admin/merchants/applications', { params }).then((r) => r.data);
}

export function auditMerchant(applicationId, approved, reason) {
  return client.post(`/admin/merchants/${applicationId}/audit`, { approved, reason }).then((r) => r.data);
}